+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 DONE
 - Settings page
   - Change username: Username taken check
+  - Delete Account functionality
 
 - SignUp page:
   - User enters email/password + firstname/lastname/username (required)
@@ -16,6 +17,7 @@ DONE
 
 - Side Navigation Bar
   - Will be the menu
+  - Contains: Dashboard, Profile ,Settings, Log out
 
 - CSS Styling
   - ProfilePage
@@ -28,5 +30,8 @@ DONE
 App.js
   - If user is not signed in, dashboard/profile/settings should redirect to /login
 
+Misc
+- DONE => ISSUE(1): Let new user take username of a deleted account when (1) signing up (2)choosing new username in settings
+
 TODO
-  - ISSUE(1): Let new user take username of a deleted account when (1) signing up (2)choosing new username in settings  
+  - Create a delete account page (i.e We're sorry to see you go! Please confirm your account deletion)